requirements.txt
blinker==1.9.0
click==8.3.2
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==26.1
psycopg2-binary==2.9.11
Werkzeug==3.1.8
